+++ src/check_jsonschema/__init__.py
@@ -11,6 +11,7 @@ def main(cli_args=None):
         disable_cache=args.no_cache,
         format_enabled=not args.disable_format,
         default_instance_filetype=args.default_filetype,
+        traceback_mode=args.traceback_mode,
     )
     checker.run()
     print("ok -- validation done")